// generated/src/app/api-models/stage-inst-hdr-sup.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class StageInstHdrSup {
  batchId: string;
  transactionId: number;
  claimStatus: string;
  processingStatus: string;
  seqMembId: number;
  memberAge: number;
  memberGender: string;
  seqPcpId: number;
  pcpIpaId: string;
  pcpType: string;
  pcpSpec: string;
  seqGroupId: number;
  planCode: string;
  eligibleEffDate: Date;
  eligibleThruDate: Date;
  lob: string;
  panelId: string;
  riderString: string;
  primaryOcExists: string;
  eligStatus: string;
  eligStatParamResult: string;
  seqProvId: number;
  seqProvAddress: number;
  seqProvContract: number;
  provParStat: string;
  provType: string;
  provSpec: string;
  provIpaId: string;
  provPcpFlag: string;
  provPanel: string;
  provIpa: string;
  seqAdmProvId: number;
  admProvParStat: string;
  admProvType: string;
  admProvSpec: string;
  admProvIpaId: string;
  seqAttProvId: number;
  attProvParStat: string;
  attProvType: string;
  attProvSpec: string;
  attProvIpaId: string;
  attProvPcp: string;
  seqVendId: number;
  seqVendAddress: number;
  submittedAuthNo: number;
  submittedSecAuthNo: string;
  authClass: string;
  authWaiveMethod: string;
  authLevel: string;
  inSvcArea: string;
  acceptMedicareAssignFlag: string;
  seqDefaultVendId: number;
  seqDefaultVendAddr: number;
  seqPaySubVendId: number;
  seqPaySubVendAddrId: number;
  authMatchOrder: number;
  seqSubsId: number;
  authClaimMatchInd: string;
  pricerBaseReimbAmt: number;
  pricerOutlierPayments: number;
  pricerAltLevelCarePaym: number;
  pricerTotalReimbAmt: number;
  pricerOutlierType: number;
  pricerAverageLos: number;
  participationFlag: string;
  parRsnCode: string;
  nonParRsnCode: string;
  bmaParamParvalue: string;
  seqMembOthCov: number;
  apiSuppressEob: string;
  batchNo: string;
  auditStatus: string;
  seqOtherProvId: number;
  groupId: string;
  pcpProvId: string;
  invalidFlags: string;
  apiProvcontract: string;
  placeOfService: string;
  readmitContinueFlag: string;
  waiveMatchOrder: string;
  privacyApplies: string;
  seqMcondId: number;
  seqOperPhysProvId: number;
  diagnosis1: string;
  diagnosis2: string;
  diagnosis3: string;
  diagnosis4: string;
  diagnosis5: string;
  diagnosis6: string;
  diagnosis7: string;
  diagnosis8: string;
  diagnosis9: string;
}